requirements.txt
streamlit
google.generativeai
python.dotenv
streamlit-chat
langchain
langchain-community
langchain-google-genai
qdrant_client
langchain-qdrant
langchain_huggingface
tiktoken
huggingface_hub
accelerate
pypdf
sentence-transformers
ctransformers
PyPDF2
chromadb
faiss-cpu